ccb_ci_script/requirements.txt
-i https://pypi.doubanio.com/simple
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
idna==2.7
nose==1.3.7
paramiko==2.4.1
prettytable==0.7.2
pyasn1==0.4.4
pycparser==2.18
pynacl==1.2.1
python-dateutil==2.7.3
rarfile==3.0
requests==2.19.1
six==1.11.0
svn==0.3.46
urllib3==1.23
